task_workers/text_similarity_worker/requirements.txt
SQLAlchemy
pymssql
beautifulsoup4
httpx
celery[redis,auth,librabbitmq]
typing
python-dotenv
jieba
sklearn
pymongo